+++ Jin/Jin/LogMgr.cpp
@@ -38,7 +38,7 @@ void LogMgr::tick(HDC _dc)
 		{
 		case LOG:
 			SetTextColor(_dc, RGB(0, 0, 0));
-			TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (*iter).Message.length());
+			TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (int)(*iter).Message.length());
 			break;
 		case WARNING:
 			SetTextColor(_dc, RGB(210, 210, 153));
@@ -48,6 +48,6 @@ void LogMgr::tick(HDC _dc)
 			break;
 		}
 
-		TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (*iter).Message.length());
+		TextOut(_dc, LT.x, LT.y + i * int(10.f * 1.5f), (*iter).Message.c_str(), (int)(*iter).Message.length());
 	}
 }
