+++ Jin/Jin/Tile.cpp
@@ -24,8 +24,8 @@ void Tile::render(HDC _dc)
 
 	if (nullptr == m_Atlas)
 	{
-		Rectangle(_dc, vRenderPos.x, vRenderPos.y,
-			vRenderPos.x + TILE_SIZE, vRenderPos.y + TILE_SIZE);
+		Rectangle(_dc, (int)vRenderPos.x, (int)vRenderPos.y,
+			(int)vRenderPos.x + TILE_SIZE, (int)vRenderPos.y + TILE_SIZE);
 	}
 	else
 	{
@@ -37,7 +37,7 @@ void Tile::render(HDC _dc)
 
 		UINT maxTileCount = maxRow * maxCol;
 
-		if (maxTileCount <= m_Idx)
+		if ((int)maxTileCount <= m_Idx)
 		{
 			return;
 		}
@@ -47,10 +47,10 @@ void Tile::render(HDC _dc)
 
 		Vec2 vLeftTop = Vec2(TILE_SIZE * iCol, TILE_SIZE * iRow);
 
-		BitBlt(_dc, vRenderPos.x, vRenderPos.y
+		BitBlt(_dc, (int)vRenderPos.x, (int)vRenderPos.y
 			, TILE_SIZE, TILE_SIZE
 			, m_Atlas->GetDC()
-			, vLeftTop.x, vLeftTop.y, SRCCOPY);
+			, (int)vLeftTop.x, (int)vLeftTop.y, SRCCOPY);
 	}
 }
